Add Company Tiancai
https://tiancai.co/

Why is it awesome?
Tiancai aims to provide a network of vetted freelancers with a stable and 100% remote work environment. Tiancai will allocate software/web/app development projects to freelancers according to their skills, so no bidding of projects is required and freelancers will get paid for every hour they work on a monthly basis. The platform is actively looking for freelance devs, UI designers and PMs.